+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Add runtime type checking to Observables and fix selector equality
 - Add runtime type checking and fix selector equality
+- *(types)* Improve observable type inference and fix Root.wrap
 ## [15.1.3] - 2025-06-11
 
 ### 💼 Other